stability: implement node liveness; first step towards new range leases
This change adds a node liveness table as a global system table.
Nodes periodically write updates to their liveness record by doing
a conditional put to the liveness table. The leader of the range
containing the node liveness table gossips the latest information
to the rest of the system.

Each node has a `NodeLiveness` object which can be used to query
the status of any other node to find out if it's live or non-live
according to the liveness threshold duration compared to the last
time it successfully heartbeat its liveness record.

The as-yet-unused `IncrementEpoch` mechanism is also added in this
PR, for eventual use with the planned epoch-based range leader leases.

Updated the range leader lease RFC to reflect current thinking.

+++ server/node_test.go
@@ -329,7 +329,7 @@ func compareNodeStatus(t *testing.T, ts *TestServer, expectedNodeStatus *status.
 	// ========================================
 	// Read NodeStatus from server and validate top-level fields.
 	// ========================================
-	nodeStatusKey := keys.NodeStatusKey(int32(ts.node.Descriptor.NodeID))
+	nodeStatusKey := keys.NodeStatusKey(ts.node.Descriptor.NodeID)
 	nodeStatus := &status.NodeStatus{}
 	if err := ts.db.GetProto(context.TODO(), nodeStatusKey, nodeStatus); err != nil {
 		t.Fatalf("%d: failure getting node status: %s", testNumber, err)
